fix(testing): WP2 — glob packages/*/tests/Contract into the Unit suite so orphaned contract tests run
phpunit.xml.dist hand-enumerated tests/Contract for exactly 7 packages, so the
other 12 (including note's NoteAccessPolicyContractTest and taxonomy's
TermAccessPolicyContractTest) silently never ran in any phpunit/composer verify
invocation — no error, no warning.

The naive glob is unsafe: phpunit.xml.dist sets failOnWarning="true", and four
abstract contract bases were named *ContractTest, so scanning them emits a fatal
"is abstract" warning. Renamed them to the repo's existing Abstract*Contract
convention (cf. queue's AbstractTransportContract) and updated every concrete
subclass's extends/use: AccessPolicyContractTest -> AbstractAccessPolicyContract
(6 subclasses), EntityStorageDriverContractTest -> AbstractEntityStorageDriverContract
(2), FieldFormatterContractTest -> AbstractFieldFormatterContract (6 in ssr),
StructuredImporterContractTest -> AbstractStructuredImporterContract. The last
carried a concrete GfmTableImporterContractTest in the SAME file (which the base
rename would have re-orphaned) — split into its own Test.php file.

With the bases renamed, the 7 lines collapse to one glob, covering future
packages by construction. New gate bin/check-contract-suite-coverage (in verify
after check-phpstan-paths) asserts the glob stays and no abstract *Test class is
reintroduced under tests/Contract. Resurrected suites all pass (221 tests across
the 12 newly-covered dirs); Unit suite 8880 -> 9092.
