+++ composer.json
@@ -494,6 +494,7 @@
         ],
         "phpstan": "phpstan analyse --memory-limit=512M",
         "check-phpstan-paths": "bash bin/check-phpstan-paths",
+        "check-contract-suite-coverage": "bash bin/check-contract-suite-coverage",
         "check-admin-coercion-patterns": "bash bin/check-admin-coercion-patterns",
         "check-admin-coercion-self-test": "bash bin/check-admin-coercion-patterns --self-test",
         "check-dead-code": "php bin/check-dead-code",
@@ -505,6 +506,7 @@
             "@cs-check",
             "@phpstan",
             "@check-phpstan-paths",
+            "@check-contract-suite-coverage",
             "@check-composer-policy",
             "@check-package-layers",
             "@check-package-layers-pl008-self-test",
@@ -527,6 +529,7 @@
         "check-admin-coercion-patterns": "Prevent regression of the useAdminConfig() migration (mission M-F). Fails when raw string-compare coercion patterns (=== '1', === '0') appear in packages/admin/app/**/*.{ts,vue} outside of configCoercion.ts, test files, or lines marked with // allow-coercion: <reason>.",
         "check-admin-coercion-self-test": "Meta-test for the admin coercion gate (audit D-5). Runs bin/check-admin-coercion-patterns --self-test: plants a synthetic violator and asserts the scanner exits non-zero, proving the gate is not a silent no-op. Wired into verify and the ci/verify-gates job so the self-test actually runs in CI.",
         "check-dead-code": "Fail-on-new dead-code gate (Phase 4 of the dead-code cleanup audit). Runs phpstan-dead-code.neon with shipmonk's detector; pre-existing baseline entries are suppressed, only NEW findings fail. See docs/audits/2026-05-17-dead-code-baseline-audit.md and bin/check-dead-code header for how to handle a new finding.",
+        "check-contract-suite-coverage": "Fail-on-regression gate for the packages/*/tests/Contract coverage invariant (WP2). Asserts phpunit.xml.dist globs packages/*/tests/Contract in the Unit testsuite (not a hand-enumerated subset), and that no abstract contract base class under those directories is named with the `*Test` suffix (which trips failOnWarning). See bin/check-contract-suite-coverage header and CHANGELOG.md for the note/taxonomy orphaned-suite history this guards against.",
         "check-getquery-bindings": "Fail-on-new gate for unbound getQuery()->execute() callsites. Baseline at tools/getquery-bindings-baseline.txt. New callsites fail CI; add to baseline with mandatory exemption comment if legitimately exempt. See CLAUDE.md § CI gates.",
         "check-dispatcher-keys": "Fail-on-new gate for provider register()/boot() methods that resolve an event dispatcher under an FQCN the production kernel-services bus does not serve (only Symfony\\Contracts\\EventDispatcher\\EventDispatcherInterface is served — see ProviderRegistryKernelServices::get()) and then addListener()/addSubscriber() with it, silently registering nothing. Baseline at tools/dispatcher-key-baseline.txt. Introduced in WP4 (framework-wide dead-subscriber sweep, audit-remediation batch 2026-07-01/02) after #1852 fixed the first live instance (relationship delete guard).",
         "check-admin-dist-fresh": "Admin dist freshness gate (D6). Fails when packages/admin/app source changes without rebuilding the committed packages/admin-surface/dist bundle, by comparing a line-ending-normalised content signature (packages/admin-surface/dist.signature) against a freshly computed one. Rebuild + refresh with bin/build-admin-dist. Prevents shipping a stale prebuilt admin SPA (e.g. the alpha.226 Edit-feedback feature that never reached the served bundle).",
