+++ packages/ssr/tests/Contract/BooleanFormatterContractTest.php
@@ -6,11 +6,11 @@ namespace Waaseyaa\SSR\Tests\Contract;
 
 use PHPUnit\Framework\Attributes\CoversNothing;
 use Waaseyaa\Field\FieldFormatterInterface;
-use Waaseyaa\Field\Tests\Contract\FieldFormatterContractTest;
+use Waaseyaa\Field\Tests\Contract\AbstractFieldFormatterContract;
 use Waaseyaa\SSR\Formatter\BooleanFormatter;
 
 #[CoversNothing]
-final class BooleanFormatterContractTest extends FieldFormatterContractTest
+final class BooleanFormatterContractTest extends AbstractFieldFormatterContract
 {
     protected function createFormatter(): FieldFormatterInterface
     {
