+++ packages/ssr/tests/Contract/DateFormatterContractTest.php
@@ -6,11 +6,11 @@ namespace Waaseyaa\SSR\Tests\Contract;
 
 use PHPUnit\Framework\Attributes\CoversNothing;
 use Waaseyaa\Field\FieldFormatterInterface;
-use Waaseyaa\Field\Tests\Contract\FieldFormatterContractTest;
+use Waaseyaa\Field\Tests\Contract\AbstractFieldFormatterContract;
 use Waaseyaa\SSR\Formatter\DateFormatter;
 
 #[CoversNothing]
-final class DateFormatterContractTest extends FieldFormatterContractTest
+final class DateFormatterContractTest extends AbstractFieldFormatterContract
 {
     protected function createFormatter(): FieldFormatterInterface
     {
