+++ packages/ssr/tests/Contract/EntityReferenceFormatterContractTest.php
@@ -6,11 +6,11 @@ namespace Waaseyaa\SSR\Tests\Contract;
 
 use PHPUnit\Framework\Attributes\CoversNothing;
 use Waaseyaa\Field\FieldFormatterInterface;
-use Waaseyaa\Field\Tests\Contract\FieldFormatterContractTest;
+use Waaseyaa\Field\Tests\Contract\AbstractFieldFormatterContract;
 use Waaseyaa\SSR\Formatter\EntityReferenceFormatter;
 
 #[CoversNothing]
-final class EntityReferenceFormatterContractTest extends FieldFormatterContractTest
+final class EntityReferenceFormatterContractTest extends AbstractFieldFormatterContract
 {
     protected function createFormatter(): FieldFormatterInterface
     {
