+++ packages/ssr/tests/Contract/HtmlFormatterContractTest.php
@@ -6,11 +6,11 @@ namespace Waaseyaa\SSR\Tests\Contract;
 
 use PHPUnit\Framework\Attributes\CoversNothing;
 use Waaseyaa\Field\FieldFormatterInterface;
-use Waaseyaa\Field\Tests\Contract\FieldFormatterContractTest;
+use Waaseyaa\Field\Tests\Contract\AbstractFieldFormatterContract;
 use Waaseyaa\SSR\Formatter\HtmlFormatter;
 
 #[CoversNothing]
-final class HtmlFormatterContractTest extends FieldFormatterContractTest
+final class HtmlFormatterContractTest extends AbstractFieldFormatterContract
 {
     protected function createFormatter(): FieldFormatterInterface
     {
