+++ packages/ssr/tests/Contract/ImageFormatterContractTest.php
@@ -6,11 +6,11 @@ namespace Waaseyaa\SSR\Tests\Contract;
 
 use PHPUnit\Framework\Attributes\CoversNothing;
 use Waaseyaa\Field\FieldFormatterInterface;
-use Waaseyaa\Field\Tests\Contract\FieldFormatterContractTest;
+use Waaseyaa\Field\Tests\Contract\AbstractFieldFormatterContract;
 use Waaseyaa\SSR\Formatter\ImageFormatter;
 
 #[CoversNothing]
-final class ImageFormatterContractTest extends FieldFormatterContractTest
+final class ImageFormatterContractTest extends AbstractFieldFormatterContract
 {
     protected function createFormatter(): FieldFormatterInterface
     {
